+++ crates/core_simd/src/first.rs
